api/requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt
tensorflow==2.8.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.8.0
matplotlib
numpy
